Allow reading arguments from file (#50)
* Allow reading arguments from file

Only allowing the user to pass credentials through command line
arguments will leave traces of the credentials in the user's history. We
can stop that from happening in two ways that comes to mind:

- put the credentials in the environment, or
- put the credentials in a file

Modifying the environment is probably one step less user friendly, since
it's easiest to modify the environment variables for all of the session,
and now every program gets access to your environment.

If you instead put the credentials in a file, Python's argparse already
has support for that out of the box. It also enables saving sets of
arguments into different files, if that's a wanted use case.

There is also a hint about referring to a file in the help text, which
required a small modification for rendering newlines properly.

Signed-off-by: Carl Helmertz <helmertz@gmail.com>

* authors: Add @chelmertz

Signed-off-by: Carl Helmertz <helmertz@gmail.com>